add confirm delete

+++ src/components/sample/agreement/AgreementDetail.vue
@@ -475,18 +475,20 @@ export default {
     },
     confirmDelete () {
       let vm = this
-      this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
-        confirmButtonText: '确定',
-        cancelButtonText: '取消',
-        type: 'warning'
-      }).then(() => {
-        vm.delete()
-      }).catch(() => {
-        this.$message({
-          type: 'info',
-          message: '已取消删除'
+      if (this.agreementForm.id && this.agreementForm.id !== '') {
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          vm.delete()
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          })
         })
-      })
+      }
     },
     new () {
       this.$emit('new')
